@@ -9,16 +9,16 @@
     "url": "https://github.com/Syncano/syncano-dashboard.git"
   },
   "engines": {
-    "node": "6.3.1",
-    "npm": "3.3.10"
+    "node": "7.0.0",
+    "npm": "4.0.1"
   },
   "dependencies": {
-    "axios": "0.15.1",
+    "axios": "0.15.2",
     "babel-polyfill": "6.16.0",
     "bluebird": "3.4.6",
-    "brace": "0.8.0",
+    "brace": "0.9.0",
     "c3": "0.4.11",
-    "card-validator": "2.3.0",
+    "card-validator": "3.0.0",
     "classnames": "2.2.5",
     "d3": "3.5.17",
     "dedent": "0.6.0",
@@ -30,38 +30,38 @@
     "js-yaml": "3.6.1",
     "json-loader": "0.5.4",
     "local-storage-fallback": "1.3.0",
-    "lodash": "4.16.4",
-    "material-ui": "0.16.0",
-    "moment": "2.15.1",
-    "moment-timezone": "0.5.6",
+    "lodash": "4.16.6",
+    "material-ui": "0.16.1",
+    "moment": "2.15.2",
+    "moment-timezone": "0.5.7",
     "mousetrap": "1.6.0",
     "normalize.css": "3.0.3",
     "numeral": "1.5.3",
-    "pluralize": "1.2.1",
+    "pluralize": "3.0.0",
     "promise-series": "1.0.0",
     "radium": "0.17.2",
-    "raven-js": "3.7.0",
-    "react": "15.3.2",
-    "react-addons-create-fragment": "15.3.2",
-    "react-addons-linked-state-mixin": "15.3.2",
-    "react-addons-pure-render-mixin": "15.3.2",
-    "react-addons-test-utils": "15.3.2",
-    "react-addons-transition-group": "15.3.2",
-    "react-addons-update": "15.3.2",
+    "raven-js": "3.8.0",
+    "react": "15.4.0-rc.3",
+    "react-addons-create-fragment": "15.4.0-rc.3",
+    "react-addons-linked-state-mixin": "15.4.0-rc.3",
+    "react-addons-pure-render-mixin": "15.4.0-rc.3",
+    "react-addons-test-utils": "15.4.0-rc.3",
+    "react-addons-transition-group": "15.4.0-rc.3",
+    "react-addons-update": "15.4.0-rc.3",
     "react-copy-to-clipboard": "4.2.3",
-    "react-dom": "15.3.2",
-    "react-dropzone": "3.6.0",
+    "react-dom": "15.4.0-rc.3",
+    "react-dropzone": "3.7.2",
     "react-fittext": "0.1.1",
     "react-helmet": "3.1.0",
     "react-inlinesvg": "github:patrykkopycinski/react-inlinesvg#build",
     "react-input-mask": "0.7.5",
     "react-responsive-grid": "0.3.4",
-    "react-router": "3.0.0-beta.1",
+    "react-router": "3.0.0",
     "react-scroll": "1.4.1",
     "react-search-input": "0.10.3",
     "react-select": "0.9.1",
     "react-stickydiv": "3.4.19",
-    "react-tap-event-plugin": "1.0.0",
+    "react-tap-event-plugin": "github:Syncano/react-tap-event-plugin",
     "reflux": "0.3.0",
     "shortid": "2.2.6",
     "sort-object": "3.0.2",
@@ -72,68 +72,78 @@
   },
   "devDependencies": {
     "autoprefixer-loader": "3.2.0",
-    "aws-sdk": "2.4.5",
-    "babel-cli": "6.16.0",
-    "babel-core": "6.17.0",
-    "babel-eslint": "7.0.0",
-    "babel-loader": "6.2.5",
+    "aws-sdk": "2.6.13",
+    "babel-cli": "6.18.0",
+    "babel-core": "6.18.2",
+    "babel-eslint": "7.1.0",
+    "babel-loader": "6.2.7",
     "babel-plugin-add-module-exports": "0.2.1",
     "babel-plugin-lodash": "3.2.9",
     "babel-plugin-transform-react-constant-elements": "6.9.1",
     "babel-plugin-transform-react-inline-elements": "6.8.0",
-    "babel-preset-es2015": "6.16.0",
+    "babel-preset-es2015": "6.18.0",
     "babel-preset-react": "6.16.0",
     "babel-preset-react-hmre": "1.1.1",
     "babel-preset-stage-0": "6.16.0",
-    "compass-mixins": "0.12.8",
+    "compass-mixins": "0.12.10",
     "css-loader": "0.25.0",
     "del": "1.2.1",
-    "eslint": "3.8.0",
+    "eslint": "3.9.1",
     "eslint-config-airbnb": "12.0.0",
     "eslint-friendly-formatter": "2.0.6",
-    "eslint-loader": "1.5.0",
+    "eslint-loader": "1.6.1",
     "eslint-plugin-import": "2.0.1",
     "eslint-plugin-jsx-a11y": "2.2.3",
-    "eslint-plugin-react": "6.4.1",
+    "eslint-plugin-react": "6.5.0",
     "express": "4.14.0",
     "file-loader": "0.9.0",
     "gulp": "3.9.1",
-    "gulp-awspublish": "2.0.2",
+    "gulp-awspublish": "3.3.0",
     "gulp-cloudfront": "0.0.14",
-    "gulp-git": "1.7.0",
-    "gulp-iconfont": "5.0.1",
-    "gulp-iconfont-css": "2.0.0",
+    "gulp-git": "1.12.0",
+    "gulp-iconfont": "8.0.1",
+    "gulp-iconfont-css": "2.1.0",
     "gulp-rename": "1.2.2",
-    "gulp-rev": "6.0.1",
+    "gulp-rev": "7.1.2",
     "gulp-rev-css-url": "0.1.0",
     "gulp-rev-replace": "0.4.3",
-    "gulp-sequence": "0.4.5",
+    "gulp-sequence": "0.4.6",
     "gulp-strip-debug": "1.1.0",
-    "gulp-task-loader": "1.3.1",
+    "gulp-task-loader": "1.4.4",
     "gulp-util": "3.0.7",
     "http-server": "0.9.0",
     "nightwatch": "0.9.8",
     "node-libs-browser": ">=0.5.3",
-    "node-sass": "3.8.0",
+    "node-sass": "3.10.1",
     "raw-loader": "0.5.1",
-    "react-prism": "3.2.1",
-    "sass-loader": "3.2.1",
-    "selenium-standalone": "5.7.2",
+    "react-prism": "4.0.0",
+    "sass-loader": "4.0.2",
+    "selenium-standalone": "5.8.0",
     "semver": "5.3.0",
     "style-loader": "0.13.1",
     "through2": "2.0.1",
     "url-loader": "0.5.7",
-    "webpack": "1.13.2",
+    "webpack": "1.13.3",
     "webpack-dev-server": "1.16.2"
+  },
+  "greenkeeper": {
+    "ignore": [
+      "del",
+      "d3",
+      "history",
+      "radium",
+      "reflux"
+    ]
   },
   "scripts": {
     "start": "gulp webpack-dev-server",
     "build": "gulp build",
     "publish": "gulp publish",
     "clean": "gulp clean",
+    "reinstall": "rm -rf node_modules && npm install",
     "lint": "eslint src --quiet --ext .js --ext .jsx --fix",
-    "lint-tests": "eslint test/e2e --quiet --ext .js --ext .jsx --fix",
-    "lint-all": "eslint src test/e2e --quiet --ext .js --ext .jsx --fix",
+    "lint-tests": "eslint test --quiet --ext .js --ext .jsx --fix",
+    "lint-all": "eslint src test --quiet --ext .js --ext .jsx --fix",
     "e2e": "./test.sh",
     "e2e-branch": "nightwatch -e chrome --suiteRetries 1 --skiptags socialLogins,demoApp,customSocket",
     "e2e-master-devel": "nightwatch -e chrome --suiteRetries 1 --skiptags demoApp,customSocket",
