@@ -6,8 +6,8 @@ import _ from 'lodash';
 
 import { FormMixin } from '../../mixins';
 
-import Store from './AuthStore';
-import Actions from './AuthActions';
+import AuthStore from './AuthStore';
+import AuthActions from './AuthActions';
 import SessionStore from '../Session/SessionStore';
 import SessionActions from '../Session/SessionActions';
 import Constants from './AuthConstants';
@@ -18,7 +18,7 @@ import AccountContainer from './AccountContainer';
 
 const AccountSignup = React.createClass({
   mixins: [
-    Reflux.connect(Store),
+    Reflux.connect(AuthStore),
     FormMixin
   ],
 
@@ -38,8 +38,8 @@ const AccountSignup = React.createClass({
     const { location, router } = this.props;
 
     if (SessionStore.isAuthenticated()) {
-      const queryNext = location.query.next || null;
       const lastInstanceName = localStorage.getItem('lastInstanceName') || null;
+      const queryNext = location.query.next || null;
 
       SessionStore
         .getConnection()
@@ -51,6 +51,11 @@ const AccountSignup = React.createClass({
             router.replace({
               pathname: queryNext,
               query: _.omit(location.query, 'next')
+            });
+          } else if (queryNext) {
+            router.replace({
+              name: 'instances',
+              query: location.query
             });
           } else {
             router.replace({
@@ -87,7 +92,7 @@ const AccountSignup = React.createClass({
   },
 
   handleSocialLogin(network) {
-    Actions.socialLogin(network);
+    AuthActions.socialLogin(network);
   },
 
   handleSuccessfullValidation(data) {
@@ -95,7 +100,7 @@ const AccountSignup = React.createClass({
 
     SessionStore.setSignUpMode();
 
-    Actions.passwordSignUp({
+    AuthActions.passwordSignUp({
       email,
       password
     });
