@@ -5,7 +5,7 @@ import { ColumnList } from '../../common/';
 
 const Column = ColumnList.Column;
 
-const DotsListItem = ({ onClickDots }) => {
+const DotsListItem = ({ onDotsClick }) => {
   const styles = {
     color: Colors.blue500,
     cursor: 'pointer',
@@ -21,7 +21,7 @@ const DotsListItem = ({ onClickDots }) => {
         data-e2e="hosting-files-dots-list-item"
       >
         <div
-          onClick={onClickDots}
+          onClick={onDotsClick}
           style={styles}
         >
           ...
