@@ -9,12 +9,11 @@ import { ColumnList } from '../../common/';
 
 const Column = ColumnList.Column;
 
-const HostingFileListItem = Radium(({ item, directoryDepth, onFolderEnter, onIconClick, showDeleteDialog }) => {
+const HostingFileListItem = Radium(({ item, onFolderEnter, onIconClick, showDeleteDialog }) => {
   const lastSlashIndex = item.path.lastIndexOf('/');
   const lastCommaIndex = item.path.lastIndexOf('.');
   const fileName = item.isFolder ? item.folderName : item.path.substring(lastSlashIndex + 1);
   const fileType = item.isFolder ? 'folder' : item.path.substring(lastCommaIndex + 1);
-  const itemPath = item.folders.slice(0, directoryDepth + 1).join('/');
 
   const fileIconConfigs = {
     html: {
@@ -71,7 +70,7 @@ const HostingFileListItem = Radium(({ item, directoryDepth, onFolderEnter, onIco
       key={item.id}
     >
       <Column.CheckIcon
-        className="col-sm-14"
+        className="col-flex-1"
         id={item.id}
         iconClassName={iconConfig.icon}
         background={iconConfig.color}
@@ -79,12 +78,6 @@ const HostingFileListItem = Radium(({ item, directoryDepth, onFolderEnter, onIco
         handleIconClick={onIconClick}
         primaryText={getFolderName()}
       />
-      <Column.Desc
-        className="col-flex-1"
-        data-e2e={`${fileName}-hosting-file-list-item-description`}
-      >
-        {`/${itemPath}`}
-      </Column.Desc>
       <Column.Desc className="col-sm-4">
         {fileSize(item.size)}
       </Column.Desc>
