@@ -1,7 +1,7 @@
 import Reflux from 'reflux';
 import _ from 'lodash';
 
-import { CheckListStoreMixin, StoreLoadingMixin, WaitForStoreMixin } from '../../mixins';
+import { CheckListStoreMixin, StoreFormMixin, StoreLoadingMixin, WaitForStoreMixin } from '../../mixins';
 
 import Actions from './HostingFilesActions';
 import SessionActions from '../Session/SessionActions';
@@ -11,17 +11,24 @@ export default Reflux.createStore({
 
   mixins: [
     CheckListStoreMixin,
+    StoreFormMixin,
     StoreLoadingMixin,
     WaitForStoreMixin
   ],
 
   getInitialState() {
     return {
-      items: [],
+      currentFolderName: '',
+      directoryDepth: 0,
+      errorResponses: [],
       filesToUpload: [],
       isLoading: true,
       isUploading: false,
-      isDeleting: false
+      isDeleting: false,
+      items: [],
+      previousFolders: [],
+      showNewFolderForm: false,
+      name: ''
     };
   },
 
@@ -46,14 +53,12 @@ export default Reflux.createStore({
 
   onCheckFolder(folder) {
     const { items } = this.data;
+    const folderToCheck = _.find(items, { id: folder.id });
     const isChecked = folder.checked;
 
-    _.forEach(items, (item) => {
-      if (_.some(folder.files, { id: item.id })) {
-        item.checked = !isChecked;
-      }
+    _.forEach(folderToCheck.files, (file) => {
+      file.checked = !isChecked;
     });
-
     this.trigger(this.data);
   },
 
@@ -63,7 +68,7 @@ export default Reflux.createStore({
     currentHostingId && Actions.fetchFiles(currentHostingId);
   },
 
-  setHostingId(hostingId) {
+  onSetHostingId(hostingId) {
     this.data.currentHostingId = hostingId;
   },
 
@@ -89,9 +94,13 @@ export default Reflux.createStore({
     this.trigger(this.data);
   },
 
-  onUploadFilesFailure() {
-    this.data.isUploading = false;
-    this.refreshData();
+  onUploadFilesFailure(uploadingStatus, response) {
+    this.data.errorResponses = [...this.data.errorResponses, response];
+    this.data.currentFileIndex = uploadingStatus.currentFileIndex;
+    this.data.lastFileIndex = uploadingStatus.lastFileIndex;
+    uploadingStatus.isFinished && removeEventListener('beforeunload', this.handleCloseOnUpload);
+
+    this.trigger(this.data);
   },
 
   onFetchFilesCompleted(data) {
@@ -109,5 +118,12 @@ export default Reflux.createStore({
     this.data.currentFileIndex = deletingStatus.currentFileIndex;
     this.data.lastFileIndex = deletingStatus.lastFileIndex;
     this.trigger(this.data);
+  },
+
+  onFinishUploading() {
+    this.data.filesToUpload = [];
+    this.data.uploadErrors = [];
+    this.data.isUploading = false;
+    this.refreshData();
   }
 });
