@@ -1,13 +1,17 @@
 import Reflux from 'reflux';
 import _ from 'lodash';
-import Actions from './ProfileActions';
-import { StoreHelpersMixin, SnackbarNotificationMixin } from '../../mixins';
 
+import ProfileActions from './ProfileActions';
+
+import { SnackbarNotificationMixin, StoreHelpersMixin } from '../../mixins';
 
 export default Reflux.createStore({
-  listenables: Actions,
+  listenables: ProfileActions,
 
-  mixins: [StoreHelpersMixin, SnackbarNotificationMixin],
+  mixins: [
+    SnackbarNotificationMixin,
+    StoreHelpersMixin
+  ],
 
   getInitialState() {
     return {
@@ -21,16 +25,12 @@ export default Reflux.createStore({
   },
 
   onFetchInvoicesCompleted(invoices) {
-    console.debug('ProfileBillingInvoicesStore::onFetchInvoicesCompleted');
-
     this.data.isLoading = false;
     this.data.invoices = invoices;
     this.trigger(this.data);
   },
 
   onFetchInvoicesFailure() {
-    console.debug('ProfileBillingInvoicesStore::onFetchInvoicesFailure');
-
     this.data.isLoading = false;
     this.trigger(this.data);
   },
@@ -40,11 +40,12 @@ export default Reflux.createStore({
       _invoice.actionDisabled = _invoice.id === invoice.id;
       return _invoice;
     });
+
     this.trigger(this.data);
   },
 
   onRetryPaymentCompleted() {
-    Actions.fetchInvoices();
+    ProfileActions.fetchInvoices();
   },
 
   onRetryPaymentFailure(message) {
@@ -52,6 +53,7 @@ export default Reflux.createStore({
       invoice.actionDisabled = false;
       return invoice;
     });
+
     this.setSnackbarNotification({ message: message.detail || message });
     this.trigger(this.data);
   }
