@@ -1,15 +1,15 @@
 import React from 'react';
-import { withRouter } from 'react-router';
 import Reflux from 'reflux';
+import { withRouter } from 'react-router';
 import _ from 'lodash';
 import numeral from 'numeral';
 import valid from 'card-validator';
 
 import { DialogMixin, FormMixin } from '../../mixins';
 
-import Store from './ProfileBillingPlanDialogStore';
-import PlanStore from './ProfileBillingPlanStore';
-import Actions from './ProfileBillingPlanDialogActions';
+import ProfileBillingPlanDialogStore from './ProfileBillingPlanDialogStore';
+import ProfileBillingPlanStore from './ProfileBillingPlanStore';
+import ProfileBillingPlanDialogActions from './ProfileBillingPlanDialogActions';
 
 import {
   FontIcon,
@@ -23,13 +23,10 @@ import {
   TableRowColumn
 } from 'material-ui';
 import { CreditCardForm, CreditCard, Dialog, Loading, Show, Slider } from '../../common/';
-import SliderSection from './SliderSection';
 
 const ProfileBillingPlanDialog = React.createClass({
-  displayName: 'ProfileBillingPlanDialog',
-
   mixins: [
-    Reflux.connect(Store),
+    Reflux.connect(ProfileBillingPlanDialogStore),
     DialogMixin,
     FormMixin
   ],
@@ -198,7 +195,7 @@ const ProfileBillingPlanDialog = React.createClass({
   },
 
   handleDismiss() {
-    Store.resetSelectedPricingPlan();
+    ProfileBillingPlanDialogStore.resetSelectedPricingPlan();
     this.handleCancel();
     if (typeof this.props.onDismiss === 'function') {
       this.props.onDismiss();
@@ -206,17 +203,24 @@ const ProfileBillingPlanDialog = React.createClass({
   },
 
   handleAddSubmit() {
-    Actions.submitPlan(this.getValidatorAttributes());
+    ProfileBillingPlanDialogActions.submitPlan(this.getValidatorAttributes());
   },
 
   handleEditSubmit() {
-    Actions.submitPlan(this.getValidatorAttributes());
+    ProfileBillingPlanDialogActions.submitPlan(this.getValidatorAttributes());
   },
 
   handleDialogShow() {
-    console.debug('ProfileBillingPlanDialog::handleDialogShow');
-    Actions.fetchBillingPlans();
-    Actions.fetchBillingCard();
+    ProfileBillingPlanDialogActions.getBillingPlans();
+    ProfileBillingPlanDialogActions.fetchBillingCard();
+  },
+
+  handleSliderChange(type, event, value) {
+    ProfileBillingPlanDialogActions.sliderChange(type, value);
+  },
+
+  handleSliderOptionClick(value, type) {
+    ProfileBillingPlanDialogActions.sliderChange(type, value);
   },
 
   renderFormNotificationsBlock() {
@@ -322,6 +326,7 @@ const ProfileBillingPlanDialog = React.createClass({
     if (!this.state.plan) {
       return true;
     }
+
     const defaultValue = 0;
     const selected = this.state[`${type}Selected`];
     let options = this.state.plan.options[type];
@@ -336,8 +341,8 @@ const ProfileBillingPlanDialog = React.createClass({
         value={typeof selected !== 'undefined' ? selected : defaultValue}
         type={type}
         legendItems={options}
-        optionClick={Actions.sliderLabelsClick}
-        onChange={Actions.sliderChange}
+        onChange={this.handleSliderChange}
+        optionClick={this.handleSliderOptionClick}
       />
     );
   },
@@ -359,7 +364,7 @@ const ProfileBillingPlanDialog = React.createClass({
 
   renderFeatures() {
     const styles = this.getStyles();
-    const features = Store.getFeatures();
+    const features = ProfileBillingPlanDialogStore.getFeatures();
 
     if (!features) {
       return null;
@@ -398,11 +403,11 @@ const ProfileBillingPlanDialog = React.createClass({
   render() {
     const styles = this.getStyles();
     const { canSubmit } = this.state;
-    const apiInfo = Store.getInfo('api');
-    const cbxInfo = Store.getInfo('cbx');
+    const apiInfo = ProfileBillingPlanDialogStore.getInfo('api');
+    const cbxInfo = ProfileBillingPlanDialogStore.getInfo('cbx');
     const sum = parseInt(apiInfo.total, 10) + parseInt(cbxInfo.total, 10);
-    const isSelectedPricingPlan = Store.isSelectedPricingPlan();
-    const selectedPricingPlanName = Store.getSelectedPricingPlanName();
+    const isSelectedPricingPlan = ProfileBillingPlanDialogStore.isSelectedPricingPlan();
+    const selectedPricingPlanName = ProfileBillingPlanDialogStore.getSelectedPricingPlanName();
     const dialogCustomActions = [
       <div style={isSelectedPricingPlan && styles.narrowWrapper}>
         <RaisedButton
@@ -411,8 +416,8 @@ const ProfileBillingPlanDialog = React.createClass({
           primary={true}
           onTouchTap={this.handleFormValidation}
           ref="submit"
-          data-e2e="confirm-button"
           disabled={!canSubmit}
+          data-e2e="confirm-button"
         />
       </div>
     ];
@@ -443,7 +448,7 @@ const ProfileBillingPlanDialog = React.createClass({
     let paymentComment = `Your monthly billing cycle will start on the 1st day of every month. Your payment for the
       current month will be prorated and charged immediately.`;
 
-    if (PlanStore.getPlanTotalValue()) {
+    if (ProfileBillingPlanStore.getPlanTotalValue()) {
       paymentComment = 'We will charge your new monthly Plan price when the next billing period starts.';
     }
 
@@ -462,35 +467,54 @@ const ProfileBillingPlanDialog = React.createClass({
             <div style={{ color: '#9B9B9B' }}>move the sliders to choose your plan</div>
           </div>
           <div style={{ paddingTop: 34 }}>
-            <SliderSection
+            <Slider.Section
               title="API calls"
               slider={this.renderSlider('api')}
               sliderSummary={apiSliderSummary}
             />
-            <SliderSection
+            <Slider.Section
               style={{ paddingTop: 50 }}
               title="Script seconds"
               slider={this.renderSlider('cbx')}
               sliderSummary={cbxSliderSummary}
             />
-
             <div className="row" style={{ marginTop: 40 }}>
               <div className="col-md-24">
                 <div style={styles.sectionTopic}>Summary</div>
                 <div style={styles.table}>
-                  <div className="row" style={styles.tableRow}>
-                    <div className="col-flex-1">API calls</div>
-                    <div className="col-md-10" style={styles.tableColumnSummary}>
+                  <div
+                    className="row"
+                    style={styles.tableRow}
+                  >
+                    <div className="col-flex-1">
+                      API calls
+                    </div>
+                    <div
+                      className="col-md-10"
+                      style={styles.tableColumnSummary}
+                    >
                       {parseInt(apiInfo.included, 10).toLocaleString()}
                     </div>
-                    <div className="col-md-10" style={styles.tableColumnSummary}>${apiInfo.total}/Month</div>
+                    <div className="col-md-10" style={styles.tableColumnSummary}>
+                      ${apiInfo.total}/Month
+                    </div>
                   </div>
                   <div className="row" style={styles.tableRow}>
-                    <div className="col-flex-1">Script seconds</div>
-                    <div className="col-md-10" style={styles.tableColumnSummary}>
+                    <div className="col-flex-1">
+                      Script seconds
+                    </div>
+                    <div
+                      className="col-md-10"
+                      style={styles.tableColumnSummary}
+                    >
                       {parseInt(cbxInfo.included, 10).toLocaleString()}
                     </div>
-                    <div className="col-md-10" style={styles.tableColumnSummary}>${cbxInfo.total}/Month</div>
+                    <div
+                      className="col-md-10"
+                      style={styles.tableColumnSummary}
+                    >
+                      ${cbxInfo.total}/Month
+                    </div>
                   </div>
                 </div>
                 <div style={{ marginTop: 30 }}>
@@ -498,10 +522,8 @@ const ProfileBillingPlanDialog = React.createClass({
                 </div>
               </div>
               <div className="col-md-11" style={{ paddingLeft: 35 }}>
-
                 <div style={styles.sectionTopic}>New plan:</div>
                 <div style={{ marginTop: 20, background: '#CBEDA5' }}>
-
                   <div style={styles.sectionTotalSummary}>
                     <div><strong>${sum}</strong>/month</div>
                     <div>+ overage</div>
