@@ -2,12 +2,13 @@ import Reflux from 'reflux';
 
 import { StoreFormMixin } from '../../mixins';
 
+import ProfileActions from './ProfileActions';
 import SessionStore from '../Session/SessionStore';
 import SessionActions from '../Session/SessionActions';
-import Actions from './ProfileActions';
 
 export default Reflux.createStore({
-  listenables: Actions,
+  listenables: ProfileActions,
+
   mixins: [StoreFormMixin],
 
   getInitialState() {
@@ -27,7 +28,6 @@ export default Reflux.createStore({
   },
 
   checkSession(Session) {
-    console.debug('ProfileSettingsStore:checkSession');
     if (Session.isReady()) {
       const user = SessionStore.getUser({});
 
