@@ -18,7 +18,7 @@ export default {
       .please()
       .create(params)
       .then((createdHosting) => {
-        if (params.isDefault) {
+        if (params.is_default) {
           return this.NewLibConnection
             .Hosting
             .please()
@@ -37,7 +37,7 @@ export default {
       .please()
       .update({ id }, params)
       .then((updatedHosting) => {
-        if (params.isDefault) {
+        if (params.is_default) {
           return this.NewLibConnection
             .Hosting
             .please()
@@ -68,6 +68,20 @@ export default {
         const hasNextFile = files.length > currentFileIndex + 1;
         const fileToUpdate = _.find(fetchedFiles, { path: file.path });
         const payload = { file: this.NewLibConnection.file(file), path: file.path };
+        const errorCallback = ({ errors, message }) => {
+          this.failure(
+            {
+              isFinished: !hasNextFile,
+              currentFileIndex,
+              lastFileIndex
+            },
+            {
+              file,
+              errors,
+              message
+            }
+          );
+        };
 
         if (fileToUpdate) {
           return this.NewLibConnection
@@ -79,7 +93,7 @@ export default {
               currentFileIndex,
               lastFileIndex
             }))
-            .catch(this.failure);
+            .catch(errorCallback);
         }
 
         return this.NewLibConnection
@@ -91,7 +105,7 @@ export default {
             currentFileIndex,
             lastFileIndex
           }))
-          .catch(this.failure);
+          .catch(errorCallback);
       });
     });
   },
