@@ -1,51 +1,40 @@
 import React from 'react';
 
-export default React.createClass({
-  displayName: 'SliderSection',
+const SliderSection = ({ style, title, slider, sliderSummary }) => {
+  const styles = {
+    title: {
+      fontSize: '1.2em'
+    },
+    slider: {
+      marginTop: 10,
+      padding: 10
+    },
+    sliderSummary: {
+      paddingLeft: 35
+    }
+  };
 
-  getStyles() {
-    return {
-      sectionTopic: {
-        fontSize: '1.3em'
-      }
-    };
-  },
-
-  render() {
-    const styles = this.getStyles();
-    const { style, title, suggestion, slider, sliderSummary } = this.props;
-
-    return (
-      <div
-        className="row"
-        style={style}
-      >
-        <div className="col-sm-8">
-          <div className="row">
-            <div
-              className="col-sm-6"
-              style={styles.sectionTopic}
-            >
-              {title}
-            </div>
-            <div
-              className="col-sm-6"
-              style={{ color: '#9B9B9B', textAlign: 'right' }}
-            >
-              {suggestion ? `suggestion based on usage: ${suggestion}` : null}
-            </div>
-          </div>
-          <div style={{ marginTop: 10, padding: 10 }}>
-            {slider}
-          </div>
+  return (
+    <div
+      className="row"
+      style={style}
+    >
+      <div className="col-md-24">
+        <div style={styles.title}>
+          {title}
         </div>
-        <div
-          className="col-sm-4"
-          style={{ paddingLeft: 35 }}
-        >
-          {sliderSummary}
+        <div style={styles.slider}>
+          {slider}
         </div>
       </div>
-    );
-  }
-});
+      <div
+        className="col-md-11"
+        style={styles.sliderSummary}
+      >
+        {sliderSummary}
+      </div>
+    </div>
+  );
+};
+
+export default SliderSection;
